refactor(jsi): move the JavaScriptCore engine layer into the shared tree
Same split as the other two engines: engine layer to jsi/jsc/ as .cpp, Apple
preamble left in a forwarding header at the original path. No Apple include
changed.

JSC's one genuine tie to Foundation was makeJSString, which built a JSString by
routing the UTF-8 through NSString to get UTF-16. Replaced with a strict
UTF-8 decoder in plain C++.

JSStringCreateWithUTF8CString is not a substitute on its own, which is
presumably why the NSString path existed: it takes a NUL-terminated C string,
so it silently truncates at an embedded U+0000 -- a legal JavaScript character.
The decoder preserves the old behaviour exactly, including falling back to
JSStringCreateWithUTF8CString on malformed input, which is what the previous
code did when NSString returned nil.

The decoder was unit-tested standalone before being wired in: 16 cases
covering ASCII, embedded NUL, 2/3/4-byte sequences, surrogate-pair splitting,
U+10FFFF, and the rejection paths (lone continuation, truncation, overlong
forms, directly-encoded surrogates, out-of-range, invalid lead bytes).

Verified: iOS TestRunner on JSC, 713 specs, 1 failure -- byte-identical to the
pre-move run, same spec (ApiTests SpecialCaseProperty_When_CustomSelector_
ImplementedInJS), which fails on this branch's parent too.

(cherry picked from commit 8a1221e6cab46319bd1ea73f4aa7b7e70d8542ed)
