+++ NativeScript/jsi/jsc/JSCHostObjects.cpp
@@ -1,4 +1,4 @@
-#include "NativeApiJSCRuntime.h"
+#include "jsi/jsc/JSCRuntime.h"
 #include "jsi/shared/NativeApiStackValueArray.h"
 
 #ifdef TARGET_ENGINE_JSC
